move IntrospectionContext to its own file
