+++ dev/com.ibm.ws.kernel.boot.nested/src/com/ibm/ws/kernel/launch/internal/FrameworkManager.java
@@ -21,12 +21,9 @@ import static org.osgi.framework.FrameworkUtil.asDictionary;
 import static org.osgi.service.condition.Condition.CONDITION_ID;
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.io.OutputStream;
 import java.io.OutputStreamWriter;
-import java.io.PrintWriter;
 import java.io.Writer;
 import java.lang.instrument.Instrumentation;
 import java.lang.reflect.Method;
@@ -79,7 +76,6 @@ import com.ibm.ws.kernel.boot.LaunchException;
 import com.ibm.ws.kernel.boot.ReturnCode;
 import com.ibm.ws.kernel.boot.cmdline.Utils;
 import com.ibm.ws.kernel.boot.internal.BootstrapConstants;
-import com.ibm.ws.kernel.boot.internal.FileUtils;
 import com.ibm.ws.kernel.boot.internal.KernelStartLevel;
 import com.ibm.ws.kernel.boot.internal.commands.JavaDumpAction;
 import com.ibm.ws.kernel.boot.internal.commands.JavaDumper;
@@ -100,7 +96,6 @@ import com.ibm.ws.kernel.productinfo.ProductInfo;
 import com.ibm.ws.kernel.productinfo.ProductInfoParseException;
 import com.ibm.ws.kernel.productinfo.ProductInfoReplaceException;
 import com.ibm.ws.kernel.provisioning.BundleRepositoryRegistry;
-import com.ibm.wsspi.logging.Introspector;
 import com.ibm.wsspi.logprovider.LogProvider;
 
 import io.openliberty.checkpoint.spi.CheckpointHook;
@@ -1272,117 +1267,6 @@ public class FrameworkManager {
             dumpedFlag.createNewFile();
         } catch (IOException e) {
             Tr.warning(tc, "warn.unableWriteFile", dumpedFlag, e.getMessage());
-        }
-    }
-
-    @SuppressWarnings("deprecation")
-    private static class IntrospectionContext {
-        private final BundleContext systemBundleCtx;
-        private final File dumpDir;
-        private int unnamedCount;
-
-        IntrospectionContext(BundleContext systemBundleCtx, File dumpDir) {
-            this.systemBundleCtx = systemBundleCtx;
-            this.dumpDir = dumpDir;
-        }
-
-        public void introspectAll() {
-            // create introspection dir in the dump dir which was created in the server's output directory
-            File introspectionDir = new File(dumpDir, BootstrapConstants.SERVER_INTROSPECTION_FOLDER_NAME);
-            if (!FileUtils.createDir(introspectionDir)) {
-                throw new IllegalStateException("introspections directory could not be created.");
-            }
-
-            try {
-                introspectIntrospectors(introspectionDir);
-                introspectIntrospectableServices(introspectionDir);
-            } catch (InvalidSyntaxException e) {
-                if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled()) {
-                    Tr.debug(tc, "Exception occured when get IntrospectableService refs: {0}", e);
-                }
-            }
-        }
-
-        private void introspectIntrospectors(File introspectionDir) throws InvalidSyntaxException {
-            Collection<ServiceReference<Introspector>> refs = this.systemBundleCtx.getServiceReferences(Introspector.class, null);
-            if (refs != null && !refs.isEmpty()) {
-                for (ServiceReference<Introspector> ref : refs) {
-                    Introspector introspector = this.systemBundleCtx.getService(ref);
-                    if (introspector != null) {
-                        try {
-                            String name = introspector.getIntrospectorName();
-                            String desc = introspector.getIntrospectorDescription();
-                            introspect(introspectionDir, name, desc, introspector, null);
-                        } finally {
-                            this.systemBundleCtx.ungetService(ref);
-                        }
-                    }
-                }
-            }
-        }
-
-        private void introspectIntrospectableServices(File introspectionDir) throws InvalidSyntaxException {
-            Collection<ServiceReference<com.ibm.wsspi.logging.IntrospectableService>> legacyRefs = systemBundleCtx.getServiceReferences(com.ibm.wsspi.logging.IntrospectableService.class,
-                                                                                                                                        null);
-            if (legacyRefs != null && !legacyRefs.isEmpty()) {
-                for (ServiceReference<com.ibm.wsspi.logging.IntrospectableService> ref : legacyRefs) {
-                    com.ibm.wsspi.logging.IntrospectableService serv = systemBundleCtx.getService(ref);
-                    if (serv != null) {
-                        try {
-                            String name = serv.getName();
-                            String desc = serv.getDescription();
-                            introspect(introspectionDir, name, desc, null, serv);
-                        } finally {
-                            systemBundleCtx.ungetService(ref);
-                        }
-                    }
-                }
-            }
-        }
-
-        private void introspect(File introspectionDir,
-                                String introspectionName,
-                                String introspectionDesc,
-                                Introspector introspector,
-                                com.ibm.wsspi.logging.IntrospectableService introspectable) {
-            if (introspectionName == null || introspectionName.isEmpty()) {
-                introspectionName = Introspector.class.getSimpleName() + '.' + unnamedCount++;
-            }
-
-            File introspectionFile = new File(introspectionDir, introspectionName + ".txt");
-
-            OutputStream out = null;
-            PrintWriter pw = null;
-            try {
-                out = new FileOutputStream(introspectionFile);
-                pw = new PrintWriter(new OutputStreamWriter(out, StandardCharsets.UTF_8));
-
-                // write header
-                if (introspectionDesc != null && !introspectionDesc.isEmpty()) {
-                    pw.println("The description of this introspector:");
-                    pw.println(introspectionDesc);
-                    pw.println();
-                    pw.flush();
-                }
-
-                // write body
-                if (introspectable != null) {
-                    introspectable.introspect(out);
-                } else {
-                    introspector.introspect(pw);
-                }
-            } catch (FileNotFoundException e) {
-                e.getCause(); // findbugs
-                Tr.error(tc, "error.fileNotFound", introspectionFile);
-            } catch (Throwable t) {
-                Tr.warning(tc, "warn.unableWriteFile", introspectionFile, t.getMessage());
-                if (out != null) {
-                    t.printStackTrace(pw);
-                }
-            } finally {
-                Utils.tryToClose(pw);
-                Utils.tryToClose(out);
-            }
         }
     }
