+++ dev/com.ibm.ws.kernel.boot.nested/src/com/ibm/ws/kernel/launch/internal/IntrospectionContext.java
@@ -1,0 +1,147 @@
+/*******************************************************************************
+ * Copyright (c) 2010, 2025 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License 2.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+package com.ibm.ws.kernel.launch.internal;
+
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
+import java.io.OutputStream;
+import java.io.OutputStreamWriter;
+import java.io.PrintWriter;
+import java.nio.charset.StandardCharsets;
+import java.util.Collection;
+
+import org.osgi.framework.BundleContext;
+import org.osgi.framework.InvalidSyntaxException;
+import org.osgi.framework.ServiceReference;
+
+import com.ibm.websphere.ras.Tr;
+import com.ibm.websphere.ras.TraceComponent;
+import com.ibm.ws.kernel.boot.cmdline.Utils;
+import com.ibm.ws.kernel.boot.internal.BootstrapConstants;
+import com.ibm.ws.kernel.boot.internal.FileUtils;
+import com.ibm.wsspi.logging.Introspector;
+
+@SuppressWarnings("deprecation")
+public class IntrospectionContext {
+
+    private static final TraceComponent tc = Tr.register(IntrospectionContext.class);
+
+    private final BundleContext systemBundleCtx;
+    private final File dumpDir;
+    private int unnamedCount;
+
+    IntrospectionContext(BundleContext systemBundleCtx, File dumpDir) {
+        this.systemBundleCtx = systemBundleCtx;
+        this.dumpDir = dumpDir;
+    }
+
+    public void introspectAll() {
+        // create introspection dir in the dump dir which was created in the server's output directory
+        File introspectionDir = new File(dumpDir, BootstrapConstants.SERVER_INTROSPECTION_FOLDER_NAME);
+        if (!FileUtils.createDir(introspectionDir)) {
+            throw new IllegalStateException("introspections directory could not be created.");
+        }
+
+        try {
+            introspectIntrospectors(introspectionDir);
+            introspectIntrospectableServices(introspectionDir);
+        } catch (InvalidSyntaxException e) {
+            if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled()) {
+                Tr.debug(tc, "Exception occured when get IntrospectableService refs: {0}", e);
+            }
+        }
+    }
+
+    private void introspectIntrospectors(File introspectionDir) throws InvalidSyntaxException {
+        Collection<ServiceReference<Introspector>> refs = this.systemBundleCtx.getServiceReferences(Introspector.class, null);
+        if (refs != null && !refs.isEmpty()) {
+            for (ServiceReference<Introspector> ref : refs) {
+                Introspector introspector = this.systemBundleCtx.getService(ref);
+                if (introspector != null) {
+                    try {
+                        String name = introspector.getIntrospectorName();
+                        String desc = introspector.getIntrospectorDescription();
+                        introspect(introspectionDir, name, desc, introspector, null);
+                    } finally {
+                        this.systemBundleCtx.ungetService(ref);
+                    }
+                }
+            }
+        }
+    }
+
+    private void introspectIntrospectableServices(File introspectionDir) throws InvalidSyntaxException {
+        Collection<ServiceReference<com.ibm.wsspi.logging.IntrospectableService>> legacyRefs = systemBundleCtx.getServiceReferences(com.ibm.wsspi.logging.IntrospectableService.class,
+                                                                                                                                    null);
+        if (legacyRefs != null && !legacyRefs.isEmpty()) {
+            for (ServiceReference<com.ibm.wsspi.logging.IntrospectableService> ref : legacyRefs) {
+                com.ibm.wsspi.logging.IntrospectableService serv = systemBundleCtx.getService(ref);
+                if (serv != null) {
+                    try {
+                        String name = serv.getName();
+                        String desc = serv.getDescription();
+                        introspect(introspectionDir, name, desc, null, serv);
+                    } finally {
+                        systemBundleCtx.ungetService(ref);
+                    }
+                }
+            }
+        }
+    }
+
+    private void introspect(File introspectionDir,
+                            String introspectionName,
+                            String introspectionDesc,
+                            Introspector introspector,
+                            com.ibm.wsspi.logging.IntrospectableService introspectable) {
+        if (introspectionName == null || introspectionName.isEmpty()) {
+            introspectionName = Introspector.class.getSimpleName() + '.' + unnamedCount++;
+        }
+
+        File introspectionFile = new File(introspectionDir, introspectionName + ".txt");
+
+        OutputStream out = null;
+        PrintWriter pw = null;
+        try {
+            out = new FileOutputStream(introspectionFile);
+            pw = new PrintWriter(new OutputStreamWriter(out, StandardCharsets.UTF_8));
+
+            // write header
+            if (introspectionDesc != null && !introspectionDesc.isEmpty()) {
+                pw.println("The description of this introspector:");
+                pw.println(introspectionDesc);
+                pw.println();
+                pw.flush();
+            }
+
+            // write body
+            if (introspectable != null) {
+                introspectable.introspect(out);
+            } else {
+                introspector.introspect(pw);
+            }
+        } catch (FileNotFoundException e) {
+            e.getCause(); // findbugs
+            Tr.error(tc, "error.fileNotFound", introspectionFile);
+        } catch (Throwable t) {
+            Tr.warning(tc, "warn.unableWriteFile", introspectionFile, t.getMessage());
+            if (out != null) {
+                t.printStackTrace(pw);
+            }
+        } finally {
+            Utils.tryToClose(pw);
+            Utils.tryToClose(out);
+        }
+    }
+}
